requirements.txt
flask==3.1.0
flask-socketio==5.5.1
mistralai==1.6.0
SpeechRecognition==3.14.1
pyautogui==0.9.54
pyperclip==1.9.0
edge-tts==7.0.2
python-dotenv==1.1.0
psutil==7.0.0
pyaudio==0.2.14
pywebview==5.4
simple-websocket==1.1.0
pyinstaller==6.12.0
